requirements.txt
attrs==23.2.0
blinker==1.8.2
click==8.1.7
flasgger==0.9.7.1
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mistune==3.0.2
packaging==24.1
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.19.1
six==1.16.0
Werkzeug==3.0.3
